refactor(brIBMr6C): move to according dir
